rename API service from Event to Weather
